clean-up script, allow to override config/variables

+++ tools/create_doc.sh
@@ -1,11 +1,29 @@
 #!/bin/bash
 
-OUTDIR=/tmp/xhtml/
-REPLACEIMAGES=
+## create_doc.sh
+##
+## wrapper around 'lv2toweb'.
+## read LV2 URIs from stdin, create documentation (incl. screenshot)
+## for each of the URIs and create index html page
+##
+## example usage:
+##   lv2ls | grep meter | grep -v _gtk | ./gendoc
 
-LV2TOWEB=../lv2toweb
-JALVGTK="GTK2_RC_FILES=/home/rgareus/src/git/ardour/build/gtk2_ardour/ardour3_ui_dark.rc LD_LIBRARY_PATH=/home/rgareus/tmp/lv2/lib /home/rgareus/tmp/lv2/bin/jalv.gtk"
-SCROT=scrot
+################################################################################
+
+### output dir ###
+: ${OUTDIR=/tmp/xhtml}
+
+### options ###
+# $KEEPOUTDIR # if set do not delete $OUTDIR
+# $REPLACEIMAGES # if not set keep existing screenshots
+
+### executables ###
+: ${LV2TOWEB=../lv2toweb}
+: ${JALVGTK="GTK2_RC_FILES=/home/rgareus/src/git/ardour/build/gtk2_ardour/ardour3_ui_dark.rc LD_LIBRARY_PATH=/home/rgareus/tmp/lv2/lib /home/rgareus/tmp/lv2/bin/jalv.gtk"}
+: ${SCROT=scrot}
+: ${JNOISE=jnoise}
+: ${JACKD=jackd}
 
 ################################################################################
 
@@ -47,7 +65,7 @@ function genscreenshot {
 function gendoc {
 	while read uri; do
 		FN=$(echo $uri | sed 's/[^a-zA-Z0-9]/_/g')
-		#genscreenshot "$uri" "$OUTDIR/${FN}.png"
+		genscreenshot "$uri" "$OUTDIR/${FN}.png"
 		$LV2TOWEB -i index.html -s "${FN}.png" "$uri" > "$OUTDIR/${FN}.html"
 		echo "$uri" >> $INDEXFN
 	done
@@ -66,29 +84,24 @@ function cleanup {
 ################################################################################
 
 # launch a dummy jack-server for jalv/screenshots
-jackd -r -s -n lv2screeny -d dummy &
+$JACKD -r -s -n lv2screeny -d dummy &
 JACKPID=$!
 sleep 2
 
 # jnoise to feed into audio-ports
-JACK_DEFAULT_SERVER=lv2screeny jnoise &
+JACK_DEFAULT_SERVER=lv2screeny $JNOISE &
 JNOISEPID=$!
 
 # used by jalv to check for external-ui's
 JALVERR=$(mktemp)
 INDEXFN=$(mktemp)
 
-trap cleanup INT QUIT TERM EXIT
-
 # TODO use xnest/xephyr for screenshots
 
-rm -rf "$OUTDIR"
+trap cleanup INT QUIT TERM EXIT
+
+test -z "$KEEPOUTDIR" && rm -rf "$OUTDIR"
 mkdir -p "$OUTDIR"
 
-echo "http://gareus.org/oss/lv2/meters#goniometer" | gendoc
-echo "http://gareus.org/oss/lv2/midifilter#keyrange" | gendoc
-#lv2ls | grep gareus | gendoc
-#lv2ls | grep gareus | grep meter | grep -v _gtk | gendoc
-# gendoc
+gendoc
 
-#cleanup
